Disable wallets instead of hide them
Fix linter

add disconnect functionality

[YLD-96] Wallet new design (#63)

* New design

* Create components for WalletDrawer and responsive view

* Complete wallet drawer

* Fix linter

Complete wallet drawer

Fix linter

Update wallet modal on select behaviours

Fix linter

Remove console log

Fix linter

Fix build

Minor improvements

Fix conflict
