+++ app/app/components/modals/Wallet/Wallet.tsx
@@ -18,7 +18,7 @@ const Wallet = ({ wallet, isSupported, onSelect, className }: WalletProps) => {
         ${
           isSupported
             ? "opacity-100 cursor-pointer hover:border-primary"
-            : "opacity-50 cursor-not-allowed hover:border-transparent"
+            : "opacity-50 cursor-pointer hover:border-transparent"
         }
          ${className}
       `}
