+++ app/app/components/modals/Wallet/WalletList.tsx
@@ -1,24 +1,23 @@
-import Wallet from "./Wallet";
-import { WalletType } from "@/utils/interfaces";
-import { useAppState } from "@/utils/context";
+import { ChainType, WalletType } from "@/utils/interfaces";
+import { WalletSection } from "./WalletSection";
 
 interface WalletListProps {
   detected: WalletType[];
   undetected: WalletType[];
-  isWalletValidForChain: (wallet: WalletType) => boolean;
-  onWalletSelect: (wallet: WalletType) => void;
-}
-
-interface WalletSectionProps {
-  title: string;
-  wallets: WalletType[];
-  isWalletValidForChain: (wallet: WalletType) => boolean;
+  isWalletValidForChain: (
+    wallet: WalletType,
+    selectedChains: ChainType[],
+  ) => boolean;
+  selectedWallet?: WalletType;
+  selectedChains: ChainType[];
   onWalletSelect: (wallet: WalletType) => void;
 }
 
 const WalletList = ({
   isWalletValidForChain,
   onWalletSelect,
+  selectedChains,
+  selectedWallet,
   detected,
   undetected,
 }: WalletListProps) => (
@@ -29,6 +28,8 @@ const WalletList = ({
     {detected.length > 0 && (
       <WalletSection
         title="Detected"
+        selectedWallet={selectedWallet}
+        selectedChains={selectedChains}
         wallets={
           detected.sort((a, b) => a.id.localeCompare(b.id)) as WalletType[]
         }
@@ -39,6 +40,8 @@ const WalletList = ({
 
     <WalletSection
       title="Other"
+      selectedWallet={selectedWallet}
+      selectedChains={selectedChains}
       wallets={
         undetected.sort((a, b) => a.id.localeCompare(b.id)) as WalletType[]
       }
@@ -47,41 +50,5 @@ const WalletList = ({
     />
   </div>
 );
-
-function WalletSection({
-  title,
-  wallets,
-  isWalletValidForChain,
-  onWalletSelect,
-}: WalletSectionProps) {
-  const { selectedWallet } = useAppState();
-  const handleWalletSelect = (wallet: WalletType) => {
-    if (wallet.isAvailable) {
-      onWalletSelect(wallet);
-    } else {
-      window.open(wallet.downloadUrl, "_blank");
-    }
-  };
-  return (
-    <div className="flex flex-col gap-4">
-      <h4 className="text-sm text-neutral-600 font-gt-america">{title}</h4>
-      <div className="grid grid-cols-2 gap-4">
-        {wallets.map((wallet) => (
-          <Wallet
-            key={wallet.id}
-            wallet={wallet}
-            isSupported={isWalletValidForChain(wallet)}
-            onSelect={() => handleWalletSelect(wallet)}
-            className={`${
-              wallet.id === selectedWallet?.id
-                ? "border-primary"
-                : "border-transparent"
-            } `}
-          />
-        ))}
-      </div>
-    </div>
-  );
-}
 
 export default WalletList;
