+++ app/hooks/index.ts
@@ -1,3 +1,2 @@
 export * from "./useWalletConnection";
-export * from "./useWalletList";
 export * from "./useContracts";
