+++ app/hooks/useContracts.ts
@@ -78,8 +78,8 @@ export function useContracts({
   }
 
   const chainKey = getChainKeyFromChain(asset?.chain);
-  const walletAddress = walletsState[chainKey].address;
-  const balance = balanceList![chainKey][assetId].balance;
+  const walletAddress = walletsState[chainKey]?.address;
+  const balance = balanceList![chainKey][assetId]?.balance;
 
   // Load token metadata
   const loadMetadata = useCallback(async () => {
